Final Push - Ready for Review
Video now does not spin unless it is being played. Javascript allows for the video playing to work only when the rotate also works. The video now rotates with the video's cage, giving a nice effect. Video now rotates with an icon, and when stopped gives a different one. Added an icon that the video starts to rotate and play, which is modified by CSS.